Fix dpdk demo env deployment
This commit fix the dpdk deployment using the demo environment.

Before this fix we had this error:

```
error: no objects passed to apply
error: error parsing STDIN: error converting YAML to JSON: yaml: line 14: did not find expected alphabetic or numeric character
```

That was because when we run the workaround for the kustomize apply we remove the `'`.

Signed-off-by: Sebastian Sch <sebassch@gmail.com>