Document the impact tests have on the cluster.
Running the tests on a production cluster may have undesired effect. Here
we try to document what all the various tests are doing to the cluster, to
let the users know in advance. General message is that CNF tests change
the "real devices" configuration and they are pretty invasive until
discovery mode lands.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>