Disable the deletion of the performance profile when running the tests.
Waiting for the deletion to succeed is time consuming.
In u/s CI the cluster is deleted after running the tests, so we don't care about cleaning.
In d/s CI we may need to run the tests multiple times, and it will be faster if the profile is not deleted.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>