Use default network for dpdk test
This PR change to dpdk test to create a policy
when the test are not in discovery mode to find and use
the default router and use that interface.

I make this change because on the default network there is always
traffic for the testpmd application to process, this makes the test stable.

Signed-off-by: Sebastian Sch <sebassch@gmail.com>