add xt_u32 feature
xt_u32 is an iptables filtering kernel module.
The module was removed from RHEL8 and brought back to life becasue there are some important CNFs that relay upon it. It will be removed again for RHEL9,
but till then we wanna have a documented way of enabling it and attesting it works.

This is basically a copy of sctp feature (minus the unblacklisting which isnt required for xt_u32).