Add a mechanism for automatically generating the detailed description of each cnf test.
We want to provide a description of the tests. At the same time, we want to take track if new undocumented tests are added.
Here we provide a way to generate (and extend) a map between the test names and their (manually fed) descriptions.
Then, we use that very same file to generate the markdown doc using a template.

Finally, when an uncommented test is found, the fill command fails. This is tied to the CI, to avoid committing new uncommented tests.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>