Don't call performance clean if discovery mode is enabled.
If performance clean function deletes the profile, even if it is the discovered one.
When running in discovery mode we want to make sure that the configuration of the cluster is not
impacted.
This could be achieved by setting the extra CLEAN_PERFORMANCE_PROFILE but it would result in a bad
ux. When setting discovery mode, user do not expect modifications to their configurations without
having to add extra flags.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>