Switch to the PAO namespace used in the doc and make the namespaces configurable.
When installing an operator, the user can choose the namespace where to install
it. The common behaviour is to use the one described in the docs, which is different
for the one being used here.

On top of that, we are making the namespaces related to the other two operators
(ptp, sriov) configurable for what is related to the tests implemented here.

Finally, the deploy part is updated to create and use the new default namespace
for PAO.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>