Disable chronyd when using ptp
This commit add a machine config to disable chronyd under the ptp kustomize.

When applying a ptp policy for time synchronization is important to disable
chronyd service so it will not try to change the machine clock.
If we don't disable it we are going to get time spikes between the phc2sys clock update
and the ntp service

Signed-off-by: Sebastian Sch <sebassch@gmail.com>