Leverage the validate test suite when waiting
Currently we use a is_ready.sh file per feature. This tampers the
"kustomize only" spirit of the deployments, plus makes things tricky
when dealing with all-in-one deployments such as ran profiles.

Here we change the waiting script to leverage the validation step of
cnf-tests. On top of that, we also extend the validation suite to check
if a tag for the dpdk imagestream is present, and to verify that the
sriov status is not in progress.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>