Add a GCP overlay for the ran profile.
Using the recently added validate only makefile rule we can check if the
manifests in a given profile are working properly, if the needed
operators are installed and running. However, the performance profile
contained in the ran profile is not suitable for cloud environments,
hence we override it with one with a selector that does not get applied.
The validation still checks the validity of the manifest and that all
the operators are deployed correctly.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>