infra: Add commit msg hook
This adds a hook to verify each commit msg include a relevant component.
Components are specified with "component: " prefix on the first line of
the commit msg (aka the subject line).

The hook include a list of components and exceptions. I trust you guyz to be careful around commits that touch any of the exception files ;-)

The commit hook can be installed by `make install-commit-hooks`. I
highly recommend any developer to do that and catch errors early.

This is all in preparation for a new ci lane that will enforce this.