cnf-tests: Control CPUs amount under latency-test pod
By having the new LATENCY_TEST_CPUS env variable, the user will be able to specify the amount of CPUs
that will be used by the pod which runs the latency tests.

The LATENCY_TEST_CPUS env variable is optional, the default value is the total amount of isolated CPUs minus one (this behavior is identical to the prior one).

Signed-off-by: Talor Itzhak <titzhak@redhat.com>