features: Bump the intel operator and switch to fec
This commit bump the intel operators to 1.3.0 by removing the custom catalog.
I also switch between the operators so it will be possible to deploy
the fec without the N3000 but not the other way around.

There is no usage of the N3000 if the fec is not deploy but
for the eAsic it's possible fine to deploy only the sriov-fec operator

Signed-off-by: Sebastian Sch <sebassch@gmail.com>