Ensure error propagates to user
When errors are detected in policy or site generation the program
returned a 0 status code indicating normal completion. This caused
errors in generating some CRs to be hidden from the user in the ArgoCD
pipeline flow. Adding in a panic() after any generated policies are
written out causes the pipeline job to fail and the logs will contain
critical debug information.

Signed-off-by: Ian Miller <imiller@redhat.com>