ztp: Add SNO accelerated startup services
These systemd services and the associated script take a set of  system
services (which would normally be restricted by a PAO performance
profile isolated cores definition), and remove any CPU restrictions at
boot and at shutdown for 10 minutes. These are windows of time when any
potential CPU starvation of customer workloads would not be an issue,
and at which time the unpinned system services allow faster reboot and
recovery times.

Signed-off-by: Jim Ramsay <jramsay@redhat.com>