infra: create a new entry in the makefile for the index build
This commit make a new entry in the makefile as other projects are using
the `setup-test-cluster` entry.

I also create an environment variable for the operator versions.
Is not possible to move it forward to 4.10 as the some operators
still don't have the 4.10 bundels in place. I will do it in a follow up PR in the future

Signed-off-by: Sebastian Sch <sebassch@gmail.com>