ztp: Support complianceType for full PGT
Give the user control over the complianceType for all CRs in a
PolicyGenTemplate and change the default to "musthave". This ensures
that, by default, we realize the CPU savings on managed clusters,
but provide a path for users to choose "mustonlyhave" policies
when needed. The difference in CPU use for managing "mustonlyhave"
policies for the current feature set is ~0.8 cpu as measured by
the prometheus query:
sum(pod:container_cpu_usage:sum{namespace=~"openshift-logging|openshift-authentication-operator|openshift-kube-controller-manager|openshift-kube-apiserver|openshift-apiserver|open-cluster-management-agent|open-cluster-management-agent-addon"})

Signed-off-by: Ian Miller <imiller@redhat.com>