ztp: Bug 2028590: Move our execstart last
There is a collision with a drop-in owned by MCO which also edits crio's
commandline by replacing ExecStart.  It turns out that the drop-in is
redundant with the environment variable that's already being set, so it
is safe for us to move our ExecStart later in the drop-in order and
simply replace the incompatible one.

Signed-off-by: Jim Ramsay <jramsay@redhat.com>