cnf-tests: Fix ovs_qos flake
Recently we're seeing 'Validate MCO removed egress/ingress'
tests flaking. From inspecting the logs the following
situation can happen in a test:

We delete the ingress MC, we see that the QoS is deleted from
the node, then we try to validate that the ingress limitation is
really deleted by creating the sender-receiver pods again.
Then we check that the traffic rate that arrives to the
receiver surpasses the ingressBitLimit that was set before.
This could be misleading, because if for some reason
the sender sends in a rate below what was set as the
ingressBitLimit, the test would fail thinking that
the ingress QoS still applies in the node.

Setting the limitMultiplier lower could significantly
reduce the chances of that happening.

Maybe TODO: redo how we check that the rate set by
the QoS is enforced/not enforced.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>