ztp: ztp-done policy: Check for the 50-nto-[worker|master] MachineConfig
This is created by NTO after PAO configures it, which may be separate in
time from when PAO creates its MachineConfig, so it's more appropriate
to ensure both MachineConfigs exist.

Unfortunately that forces us to have different policies depending on
whether the performance profile is for master or worker nodes, so we
need to differentiate this at the group level.

Signed-off-by: Jim Ramsay <jramsay@redhat.com>