features: Update PTP config to pass validations
The ptp-operator has a new validating webhook which ensures that the
parameters in the PTP config are valid. The parameters used in CI don't
conform so CI fails. In specific the summary_interval and
logSyncInterval need to be set to the same value. The CI is being
updated to use -4 to match the "right" setting when "fast event
notification" is being used.
Validator:
https://github.com/openshift/ptp-operator/blob/master/api/v1/ptpconfig_webhook.go#L122-L124

Signed-off-by: Ian Miller <imiller@redhat.com>