ztp: Support existing customer Storage config
Existing customers use StorageLocalVolume rather than StorageLV. Add
symlink to allow either naming for this file for backwards
compatibility.

Some customers want to annotate a StorageClass as the default class. We
need to provide the StorageClass CR to allow this annotation (rather
than relying on its implicit creation along with the LocalVolume).

Signed-off-by: Ian Miller <imiller@redhat.com>