ztp: control OVS worker threads on DU profile
The ovs-vswitchd process on DU-configured SNO runs as part of the
management overhead set of processes which are constrained to the
management cores. When ovs-vswitchd starts up, it samples the system for
number of cores and the cpu-affinity mask to determine how many worker
threads should be created.  That sampling can happen before we constrain
ovs to the set of management cores, which results in OVS still using a
large number of worker threads even after we constrain it to the smaller
set of management cores.

Re-trigger OVS to recalculate the worker threads using the ovs-vsctl
command.  We use the same calculation done by OVS to re-configure it so
that the number of worker threads matches the reduced set of cores.