ztp: Redo the ContainerFile for ztp-site-generator
This removes the reliance on an external tarball so the Prow CI image
builder can build this for us.

It also simplifies the build process a lot, relying on the existing
makefiles in our source control to set up the kustomize plugin directory
structure rather than duplicating it, and removing a lot of extra
copying.

Signed-off-by: Jim Ramsay <jramsay@redhat.com>