ztp: Suppress NMStateConfig when nodeNetwork empty
If the user does not supply any networking data in the nodeNetwork field
for a given node we should not generate the NMStateConfig CR. The
resultant NMStateConfig CR would be empty (no config/interfaces) and
causes the install to not complete (crash noted in infraenv). This patch
detects the empty nodeNetwork configuration in the user supplied
SiteConfig and suppresses the generation of NMStateConfig for that node.

Signed-off-by: Ian Miller <imiller@redhat.com>