ztp: bug 2002628: gracefully reconcile deleted policies
Removing one of the policies from the PGT does
not lead to removal of the generated policy from the
hub and does not remove the related objects from the spoke
This commit adds a comparison between the generated policies
and the policies present on the hub. If some of the
policies have been removed from the PGT list but still configured
on the hub, they are being deleted.

Signed-off-by: Vitaly Grinberg <vgrinber@redhat.com>