ztp: Bug 2047291: fix siteConfig test cases which are dependant on MachineConfigs in ztp/source-crs/extra-manifests
Several siteConfig test cases have expected MachineConfig (MC) results that are currently being tested against the MachineConfigs in ztp/source-crs/extra-manifests. These MCs are dynamically rendered in ztp/extra-manifests-builder. Thus, if a change to the underlying scripts / configs that are used in the MCs is made, it is required to re-generate the new MCs and commit these new MCs with the code changes in order to ensure that the test cases pass.

This commit addresses the above problem by decoupling the actual extra-manifests in ztp/source-crs/extra-manifests by introducing an independent, smaller set of sample MachineConfigs in a segregated ztp/siteconfig-generator/siteconfig/testdata/ area.

The following MCs have been applied for the siteConfig test cases:
- One pair of static MachineConfig (worker/master)
- One templated MachineConfig (will be instantiated for both master / worker)
- One copy of the WorkloadPartitioning MCs for SNO cases.