ztp: Fix image registry partition template on multi node cluster
If the SiteConfig contains multiple nodes per cluster, the siteconfig
generator fails as it tries to create the same manifest for each one
of the nodes.
Since the config for the registry should be the same for all the nodes
in the cluster, this patch removes the range on all clusters in the
template and it only consumes the SiteConfig data of the first node.