cnf-tests: verify minimal CPUs requirement for oslat test
We should handle the amount of minimal CPUs for oslat test more carefully.
When HT disabled oslat needs 2 CPUs.

But when HT enabled oslat needs at least 4 CPUs:
One CPU should be used to run oslat's main thread.
The second is the sibling of the first one, which should be excluded from the list of the tested CPUs,
Because it might cause to false spikes (noisy-neighbor issue).
The third one is the actual CPU to be tested, but due to SMT alignment restriction we need its sibling too.
Four in total.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>