infra: Improve feature-wait timeout
The current timeout can result a wait of ~5 hours
if one of the validation tests is failing.
In a working run on BM the max time is takes for
all the CNF features to be ready is 50 minutes.
We can reduce the timeout to 90 minutes and allow
setting the timeout variable.