ztp: Revert "ztp: Set summary_interval and logSyncInterval to 0"
This reverts commit 3633c372dc6952715fe522cfdedb040f7dfb6cdc.

The underlying reason to align the summary_interval and logSyncInterval
was due to a validating webhook in the PTP operator. This validation has
been removed and the values no longer need to be aligned. This commit is
being reverted because the change to the source CRs would cause an
update to the customer's configurations on upgrade which is no longer
required. To avoid this update and potential disruption the change is
being reverted. The customer can supply whatever values are appropriate
for their use case in the PolicyGenTemplate override.