cnf-tests: Split setup and cleanup by feature
Doing this for parallel run of the tests.
To be able running more than one instance of
cnf-tests at a time, by feature, we need to
split the setup and cleanup for every tested
feature, so a running instance of the tests
won't delete the resources of another tested
feature.