ztp: Enable workload partitioning on SNO worker
SNO expansion scenario occurs when initially a SNO
is deployed into production, and at a later date the
cluster size is increased by the addition of a worker node.
The resulting cluster will contain two nodes:
1. Original SNO serving as the control plane for itself and the additional node
2. One additional node without any control plane components (controlled by the SNO).
The additional worker is running the same type of workload as the original SNO,
so all the DU related configurations and optimizations apply to the worker
node as well. This includes workload partitioning.

This commit allows generation of workload partitioning
manifests on the additional worker node described above.

Signed-off-by: Vitaly <vgrinber@redhat.com>