ztp: Prevent changing immutable objects on sno exp
This change allows deploying additional SNO workers with ZTP.
When adding a worker to an existing SNO, user
adds an additional node to the existing SiteConfig.
The Generator is processing the modified siteconfig
and adds an additional line in AgentClusterInstall:
workerAgents: 1
However, agentclusterinstall is immutable.There is an
admission webhook that declines the change,
and the sync fails:
Attempted to change AgentClusterInstall.Spec which is
immutable after install started, except for ClusterMetadata
fields. Unsupported change: ProvisionRequirements.WorkerAgents: (0 => 1)