ztp: run PTP daemons on worker nodes only
ZTP DU workflow provisions PtpOperatorConfig resource in case there
is a need for ptpEventConfig setting, and for standard clusters regardless
of ptp event needs.
For SNOs, the PtpOperatorConfig resource includes daemonsetNodeSelector
set to "master".
If at a later time SNO is expanded with one or more workers, PTP operator
would not create linuxptp-daemon containers on the worker node(s). Any
attempt to change the daemonsetNodeSelector will result in ptp daemon
restart and time synchronization loss.
This commit sets daemonNodeSelector to "worker" unconditionally. It fits
all the deployment types known so far (both SNO and 3NC nodes have the
"worker" label, and will be selected. For the standard clusters we only
want to deploy PTP on workers)