ztp: OCPBUGS-239: run SR-IOV daemons on workers
ZTP DU workflow provisions SriovOperatorConfig resource.
For SNOs, the SriovOperatorConfig resource includes
configDaemonNodeSelector set to "master".
If at a later time SNO is expanded with one or more workers,
SRIOV operator would not create sriov-device-plugin and
sriov-network-config-daemon pods on the worker node(s).
Any attempt to change the configDaemonNodeSelector will
result in create sriov-device-plugin and
sriov-network-config-daemon pods restart and possible
network connectivity loss.

This commit sets configDaemonNodeSelector to "worker"
unconditionally. This fits all the deployment types known so far
(both SNO and 3NC nodes have the "worker" label, and therefore will
be selected, and on the standard clusters we only want to deploy
the daemons on workers)

Signed-off-by: Vitaly Grinberg <vgrinber@redhat.com>