cnf-tests: Allow focusing on features by ginkgo describe or context
By adding the brackets we can focus on features by
ginkgo describe or context. Some tests have in their
It or By the name of another feature, different than
the one we currently want to test (for example
integration tests between features). This way we will
be able to run only tests related to specific feature.