ztp: OCPBUGS-217: run PTP daemons on worker nodes only - revert
Reverts #1213
The reason for creating the original PR was to make the PTP
daemon selector suitable for deployments besides SNO. This PR is
being reverted because the change to the source CRs would cause a
potentially disruptive update to the customer's cluster upon upgrade.
A better way to tackle the original problem would be updating
the reference PolicyGenTemplate.