ztp: OCPBUGS-503: run SR-IOV daemons on workers - revert
Reverts #1228
The reason for creating the original PR was to make the SR-IOV
daemon selector suitable for deployments besides SNO. This PR is
being reverted because the change to the source CRs would cause a
potentially disruptive update to the customer's cluster upon upgrade.
A better way to tackle the original problem would be updating
the reference PolicyGenTemplate.