ztp: OCPBUGS-239: set SR-IOV node selector
This commit provides reference for overriding SR-IOV
configDaemonNodeSelector to "worker" instead of the default
"master" predefined in the source-crs. The change is recommended on SNOs
for forward compatibility with possible SNO expansion.
When the default setting is left intact, then in case of an SNO expansion
with one or more workers, SR-IOV operator would not create daemon containers
on the worker node(s). Any attempt to change the configDaemonNodeSelector
will result in daemon restart and possible downtime.