cnf-tests: intentionally ignore NROP setup errors
We need to ignore the setup error, if any, in the NROP e2e
suite. Improving in this area calls for a major overhaul
in the NROP e2e suite, which is not ready yet.

The suite is equipped to handle re-discover later in the stage
anyway, so this change won't hurt runs of the suite itself, but
it will unlock *other* suites to run, fixing instances of this error:

```
/remote-source/app/cnf-tests/testsuites/e2esuite/test_suite_test.go:106
   Unexpected error:
      <*meta.NoKindMatchError | 0xc001c6c1c0>: {
          GroupKind: {
              Group: "topology.node.k8s.io",
              Kind: "NodeResourceTopology",
          },
          SearchedVersions: ["v1alpha1"],
      }
      no matches for kind "NodeResourceTopology" in version "topology.node.k8s.io/v1alpha1"
  occurred
  /remote-source/app/cnf-tests/testsuites/e2esuite/test_suite_test.go:109
```

Signed-off-by: Francesco Romani <fromani@redhat.com>