cnf-tests: run cnf-tests on properly isolated cpus
We need to replace the `exec.Command` with `execve` syscall. `exec.Command` is using fork, which means that it leaves the parent process alive, and the parent process affined to the entire CPU set which might cause latency spikes.
`execve` OTOH is replacing the parent (the runner process) with the actual latency test process (and essentially removing the parent process).
This way we guarantee that no other process is running on the tested CPU set beside the test's process itself.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>