cnf-tests: Filter tests by features suites
Adding brackets to focus on tests by feature.
If FEATURES var is used for ginkgo focus we
should do the filtering by suites ([feature]).
Without the brackets we may run tests that have
a feature name in their description while they
belong to a feature suite that we don't want to
currently test.