vendor: bump NTO dependecies 20221113
This bump contains the changes from openshift/cluster-node-tuning-operator#499,
which modifies the latency test to read the log from the cnf-tests pod itself instead from a log file.

Hence, it needs to be part of this PR, because from now on the latency measurement tools will log their output to stdout alone.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>