cnf-tests: add PSA for performance tests` NS
In case the performance testing namespace got created
on a different context (IOW, it already exists)
We might not have the PSA set for the NS.

Let's add them to make sure the testing are running properly.

The risk is minimal, since this is a controled environment and
the namespace will get deleted at the end of the test anyway.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>