ztp: OCPBUGS-4449: Disable ice module when doing a kdump
Adding a MachineConfig to remove the ice module from kdump to prevent
kdump failures on certain servers. This is a temporary workaround for
RHELPLAN-138236 and can be removed when that issue is fixed.

Note: This is done by editing the kdump config file. When this
MachineConfig is removed, the changes to the kdump config file will
not be undone. This is OK because the ice module is not used when
doing a kdump.

Signed-off-by: Bart Wensley <bwensley@redhat.com>