ztp: example config of composable Openshift
installConfigOverrides is the generic way of passing install-config
parameters through the siteConfig.
This change adds a sample installConfigOverrides that configures
the composable openshift feature with the 'capabilities' field.  It
removes all but the marketplace component from the optional set of
components.  The resulting 'capabilities' in the install-config will be:

capabilities:
  baselineCapabilitySet: None
    additionalEnabledCapabilities:
      - marketplace