ztp: remove systemd from accelerated-container-startup
OCPBUGS-5380

Tests showed that un-restricting systemd does not yield much boot time
savings.  It also interferes with how many handler threads ovs starts
with.  If accelerated-startup.service starts before the
ovs-vswitchd.service, then this script will un-restrict systemd and when
ovs starts, it will inherit the affinity of systemd which is the full
set of cores.