ztp: Remove incorrect macAddress field
The macAddress field under
spec.clusters[].nodes[].nodeNetwork.config.interfaces[] is not needed.
The network configuration is matched to interfaces based on the
macAddress under the spec.clusters[].nodes[].nodeNetwork.interfaces[].
In addition, if the config.interfaces field is included it needs to be
mac-addresses. Since it is not necessary this PR simply removes it.

Signed-off-by: Ian Miller <imiller@redhat.com>