ztp: Comment out workloadHints in ZTP 4.13
A new workloadHints spec was added to 4.13 PerformanceProfile source-crs which does not exist in previous ZTP versions. This results in an installation failure for earlier releases.

This commit comments out the workloadHints object and adds it to the example files. When ZTP becomes version-aware, the workloadHints object will be uncommented.

Signed-off-by: Sharat Akhoury <sakhoury@redhat.com>