cnf-tests: Add JUnit merger
This commit adds a simple JUnit merger program.
The program gets as input file names of JUnit files generated
by Ginkgo, and merges them together and outputs to the file
name as the "output" flag (default is stdout).
This will allow us to preserve the same JUnit output as before the split.

This program includes unit tests.

Signed-off-by: liornoy <lnoy@redhat.com>