cnf-tests: Build just latency tests in Dockerfile
The cnf-tests split implementation breaks the cnf-tests container.
We don't build the test binaries anymore nor run the functional tests in
a containerized way.

cnf-tests container is a supported product for running latency tests,
so we need to rewrite the cnf-tests container Dockerfiles to include
only the latency test.

Additionally it would be a god time to improve user interface to allow
users
to easily invoke the tests without providing many different parameters
and environment variables,
thus providing better UX and preventing errors.

Signed-off-by: Jose Luis Ojosnegros Manchón <jojosneg@redhat.com>