ztp: Set Ironic inspection to enabled by default
ACM 2.8 enables the converged flow by default. In the converged flow, the Assisted Service
BareMetal Agent Controller (BMAC) does not synchronize the inventory with the Baremetal
Host objects and this task is now handled by Ironic.

This update sets the Ironic inspection as enabled by default. By doing so, the inspection
process will be automatically initiated for BareMetalHosts (BMHs) during deployment.
As a result, the hardware inventory will be populated.

Additionally, the relevant test cases have been updated to reflect this change in the
default setting.

Signed-off-by: Tao Liu <tali@redhat.com>