ztp: Include SR-IOV-related kernel command-line arguments on day-0
This update ensures that the day–0 extra manifest incorporates the necessary
SR-IOV-related kernel command-line arguments. By introducing a Machine
Configuration (MC), we address the issue of the sriov-network-config-daemon
triggering a node reboot after applying the site policy on day-2.

This is necessary to address the potential race condition where the site policy
is applied prior to MCP rebooting the node after applying the performance profile.

Signed-off-by: Tao Liu <tali@redhat.com>