cnf-tests: use pod affinity in `tuningcni` tests
Tuning CNI test cases leverage multus macvlan networks, so
pods must be located in the same host to ping each other.

Previuous way to ensure pods are co-located was not good, as the field
`pod.Spec.NodeName` is not populated when the pod is created.

Use `pod.Spec.Affinity.podAffinity` using pod labels to ensure
client and server pods run on the same host.

Signed-off-by: Andrea Panattoni <apanatto@redhat.com>