cnf-tests: Modify the test entrypoint scripts
This removes the support of running the cnf-tests in a container.
We change flags' names because we now use Ginkgo tool and not `go test`.

Privously we had three test suites: validationsuite,
configsuite, and cnftests. Those test suites would compile external
suites together (e.g. the cnftests complied together metallb, sriov, nto
and ptp), where now we invoke the external suites directly from
their clone in the submodule dir.

In the run-functests.sh script we need to map the suite + external suite
to the correct path to invoke the test from. This is handled by the TESTS_PATHS
var. We also take care of producing the same JUnit report as before.

Before running each test suite we print the commit it references, e.g.:
"now testing validationsuite integration: 7bb53ae - cnf-tests: Build just
latency tests in Dockerfile"

Note that the EXTERNAL_SUITES is an optional var to run a subset of
external test suites.

Finally, we make the k8s-reporter to ignore errors of "file already
exists" when creating the reporter dir.

Signed-off-by: Lior Noy <lnoy@redhat.com>