cnf-tests: fix `ApplyKubeletConfigToNode` cleanup
When `ApplyKubeletConfigToNode` moves the node back to
the previous role, we have to wait for the relative MachineConfigPool to
update the node. Since the function has no clue about which
MCP the node belonged to before applying the KubeletConfig, the best
strategy is to wait for at least one MCP to start updating and then wait
for the whole cluster (every MCP) to be stable.

Before these changes, the test case that runs after the cleanup function
is likely to fail, as it found a cluster in a non-stable state.

Introduce utility functions `WaitForAllMCPStable()`
and `WaitForAtLeastOneMCPUpdating()` to query the state of the cluster
without specifying a single MCP.

Signed-off-by: Andrea Panattoni <apanatto@redhat.com>