cnf-tests: Dump correct code line on failure
Custom ginkgo FailHanlder adds an item in the call stack
when handling a failure and ginkgo uses `panic(...)` to get the
stacktrace and to infer the user code location of the failing assertion.

This commit ensure the `callerSkip ...int` parameter is incremented.

Prior to this commit, failing messages appeared as

```
In [It] at: /tmp/cnf-jc658/cnf-features-deploy/vendor/github.com/onsi/gomega/internal/assertion.go:62 @ 09/11/23 20:31:50.623
```

Signed-off-by: Andrea Panattoni <apanatto@redhat.com>