cnf-tests: Fix the test-bin target
Currently, the "test-bin" target invokes the "init-git-submodules".
This is, in his turn, invoke "fetch -all" which creates connection
to the public github repo. However, in Red Hat's mid-stream container
image builder, we are creating the image in a disconnected environment.

So here we remove the call to the "init-git-submodules" target,
and instead only invoke "git submodule update --init --force".

Signed-off-by: Lior Noy <lnoy@redhat.com>