cnf-tests: Set DISCOVERY_MODE to true by default
Modify the 'test-run.sh' script, which serves as the entry point for
the latency tests, and set the DISCOVERY_MODE variable to true.
This adjustment eliminates the need for users to set it manually.
However, users still have the option to override the variable and
set it to 'false' if desired.

Signed-off-by: Lior Noy <lnoy@redhat.com>