cnf-tests: Fixes snyk file content
Currently the SAST workflow that we are adding to cnf-features-deploy repo fails for the following reason:
Could not parse ignore rules to glob { path: '/go/src/github.com/openshift-kni/cnf-features-deploy/.snyk' }
Please make sure ignore file follows correct syntax
The failure occurs in "snyk code test" command that is used in the security workflow.
By changing the current .snyk file content to a new one the parsing of the file is successful and so is the security scanning.