cnf-tests: Separate ZTP and CNF-Tests CI Jobs
This commit modifies the Makefile, creating a distinct Makefile
target for the ZTP CI job. This adjustment aims to prevent conflicts
in case of issues on the ZTP side, effectively blocking merges for
the CNF-Tests component.

Alongside this commit, we will introduce a new Prow job
in the openshift/release repository. This job will activate the
new target but exclude it for changes exclusive to CNF-Tests.

Signed-off-by: Lior Noy <lnoy@redhat.com>