ztp: Relocate the IBU source-cr
The IBU source-cr is relocated to the ibu directory. The seed image reference is also commented out to allow for concise stage transitions. The ManagedClusterSetBinding name is modified to reflect the matching clusterSet.

The PR also introduces a new IBU status CR to be used for checking the status of the various IBU stages.