cnf-tests: Read guaranteed CPU list for cgroupv2
Guaranteed workload pod can be in a different cgroup
if on the node there have been applied a PerformanceProfile.

The Cluster Node Tuning Operator configures the cpuset
cgroup on specific condition and it doesn't come back to
the original configuration when the node is moved to a differenent
MachineConfigPool.

Update the logic that DPDK test cases use to retrieve the list of
dedicated CPUs.

Refs:
https://github.com/openshift/cluster-node-tuning-operator/blob/a4c70abb71036341dfaf0cac30dab0d166e55cbd/assets/performanceprofile/scripts/cpuset-configure.sh#L9
Signed-off-by: Andrea Panattoni <apanatto@redhat.com>