ztp: OCPBUGS-37572: Fix aZTP for SiteConfig v1
Accelerated ZTP relies on ZTP adding a second ConfigMap reference in the
AgentClusterInstall CR.  However, SiteConfig v1 only supports adding a
single ConfigMap reference through the manifestsConfigMapRef field of
the AgentClusterInstall CR.
This fix changes manifestsConfigMapRef to manifestsConfigMapRefs and
allows adding the second ConfigMap reference.  It only does that if
SiteConfig v1 is used and the "accelerated-ztp" label is found on the
SiteConfig to be backward compatible.