ztp: Ensure BMH deletion before InfrEnv and NS
OCPBUGS-42945, during node or cluster cleanup the BMH must complete
deletion before the InfraEnv, NMStateConfig or Namespace are removed.
This means the BMH must be in a higher wave (later for creation, earlier
for deletion) than those other CRs.

The requirement comes from the cleanup phase of assisted
installer/baremetal operator. If the BMH automatedCleanupMode is not
disabled the node is rebooted into a new ISO image which cleans the
disk. The creation of this ISO requires the Namespace to be not in the
deleted state so that a new PreProvisioningImage CR can be created. The
ISO creation also requires that the correct NMStateConfig CR be present
(it is included in the ISO). The booting of the ISO requires the
InfraEnv CR to be present.

Signed-off-by: Ian Miller <imiller@redhat.com>